+++ CHANGELOG.md
@@ -58,11 +58,18 @@ to be rendered more efficiently client-side.
 
 Some functionality that was previously deprecated has been removed:
 
-  * Implicit assigns in `live_component` do-blocks is no longer supported
+  - Implicit assigns in `live_component` do-blocks is no longer supported
+  - Passing a `@socket` to `live_component` will raise when possible
 
 ### Enhancements
 
+  - Add `<.live_component module={FormComponent} id="form" />`
+
 ### Bug fixes
+
+### Deprecations
+
+  - `<%= live_component MyModule, id: @user.id, user: @user %>` is deprecated in favor of `<.live_component module={MyModule} id={@user.id} user={@user} />`. Notice the new API requires using HEEx templates. This change allows us to further improve LiveComponent and bring new features such as slots to them.
 
 ## 0.16.4 (2021-09-22)